In the Release Notes:
I updated the wording of one feature as follows:
Users can now click and drag a session tab to undock it from the main Connection Manager window.
The tab becomes a standalone window that users can move to a different monitor or to a different position on their desktop.
Users can also re-dock the window as a tab next to any of the other tabs in the main Connection Manager window.

I also added a Known Issues section, with this one known issue in it:
Clicking on “X” to close an undocked session window, then immediately trying to move the window back to the docking area could lock the other tabs.
Workaround: Close and re-start Connection Manager.